backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
scikit-learn==1.6.1
numpy==2.2.3
pandas==2.2.3
joblib==1.4.2
python-multipart==0.0.20
tensorflow
Pillow
